pyproject.toml
[tool.poetry]
name = "breach-search"
version = "0.1.0"
description = "Breach PII search tool using Azure AI Search and fuzzy matching"
authors = ["PwC"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
rapidfuzz = "^3.0"
azure-search-documents = "^11.4"
openpyxl = "^3.1"
xlrd = "^2.0"
xlwt = "^1.3"
pydantic-settings = "^2.0"
python-dotenv = "^1.0"
pyyaml = "^6.0"
click = "^8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-mock = "^3.12"

[tool.poetry.scripts]
breach-search = "app.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
